requirements.txt
torchdiffeq
torchsde
cython==0.29.33
pandas==1.5.3
numpy==1.24.3
h5py==3.8.0
dill==0.3.6
matplotlib
opencv-python
pyrebase4
openpyxl
networkx
addict==2.4.0
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cycler==0.11.0
ffmpeg==1.4
fonttools==4.38.0
idna==3.4
imageio==2.26.1
importlib-metadata==6.1.0
joblib==1.2.0
kiwisolver==1.4.4
markdown==3.4.2
markdown-it-py==2.2.0
matplotlib==3.5.3
mdurl==0.1.2
mmcv-full
model-index==0.1.11
networkx==2.6.3
ninja==1.10.2.3
opencv-python==4.5.5.62
openmim==0.3.7
ordered-set==4.1.0
overrides==7.3.1
packaging==23.0
pillow==9.4.0
protobuf==3.20.3
pycocotools==2.0.6
pycparser==2.21
pygments==2.14.0
pyparsing==3.0.9
pytorch-metric-learning==2.1.1
pywavelets==1.3.0
pyyaml==6.0
requests==2.28.2
rich==13.3.2
scipy
tabulate==0.9.0
tensorboardx==2.6
threadpoolctl==3.1.0
tifffile==2021.11.2
typing-extensions==4.5.0
urllib3==1.26.15
yacs==0.1.8
yapf==0.32.0
zipp==3.15.0
easydict==1.9
pytz==2022.7.1
six==1.16.0
tqdm==4.65.0
msgpack-python==0.5.6
wandb
loguru
tiktoken
peft==0.14.0